requirements.txt
# Django y REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
drf-spectacular==0.26.5

# Base de datos
psycopg2-binary==2.9.7

# Cache y tareas asíncronas
redis==5.0.1
celery==5.3.4

# Autenticación JWT
djangorestframework-simplejwt==5.3.0

# Generación de PDFs y QR
reportlab==4.0.7
qrcode[pil]==7.4.2
Pillow==10.1.0

# Utilidades
python-decouple==3.8
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
hypothesis==6.88.1
factory-boy==3.3.0